+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Adding the `clustering/record-linkage/leader` namespace.
 * Adding the `keyers/name-power-set` namespace.
 * Adding the `keyers/normalize` namespace.
+* Adding the `tokenizers/fingerprint/name` namespace.
 * Adding the `tokenizers/skipgrams` namespace.
 * Fixing & optimizing `clustering/k-means`.
 * Dropping the `helpers/random` namespace inf favor of [pandemonium](https://github.com/Yomguithereal/pandemonium).